+++ modules/desktop/electron/libs/ipc.ts
@@ -1,21 +1,17 @@
 import { ipcMain, app, BrowserWindow } from "electron";
-import { createReadStream, statSync } from "fs";
 import { getInstalledPackages } from "./tea-dir";
 import { readSessionData, writeSessionData } from "./auth";
 import type { Packages, Session } from "../../src/libs/types";
 import * as log from "electron-log";
-import { post } from "./v1-client";
-import { deepReadDir, deletePackageFolder } from "./tea-dir";
-import path from "path";
-
+import { syncLogsAt } from "./v1-client";
 import { installPackage, openTerminal, syncPantry } from "./cli";
 
 import initializeTeaCli from "./initialize";
 
 import { getUpdater } from "./auto-updater";
-import fetch from "node-fetch";
 
 import { loadPackageCache, writePackageCache } from "./package";
+import { nanoid } from "nanoid";
 let teaProtocolPath = ""; // this should be empty string
 
 export const setProtocolPath = (path: string) => {
@@ -106,38 +102,19 @@ export default function initializeHandlers() {
 	ipcMain.handle("submit-logs", async () => {
 		try {
 			log.info("syncing logs");
-			const logDir = path.join(app.getPath("home"), "Library/Logs/tea");
-			// ['/Users/neil/Library/Logs/tea/main.log']
-			const logFiles = await deepReadDir({ dir: logDir });
-			let deviceId = "";
-			const files = logFiles.map((p) => {
-				const paths = p.split("/");
-				deviceId = paths[2]; // temp hack use developers username instead of gui_id
-				return paths.pop();
-			});
+			const { device_id } = await readSessionData();
+			const logId = [device_id, nanoid()].join("---");
 
-			const signedUrls = await post<{ [key: string]: string }>(`/gui/${deviceId}/sync-log-files`, {
-				files
-			});
-			if (signedUrls) {
-				for (const key in signedUrls) {
-					const fileIndex = files.indexOf(key);
-					const filePath = logFiles[fileIndex];
-					if (filePath) {
-						const payload = createReadStream(filePath);
-						const response = await fetch(signedUrls[key], {
-							method: "PUT",
-							body: payload,
-							headers: {
-								"Content-Length": statSync(filePath).size.toString()
-							}
-						});
-						log.info("uploading log:", key, response.status);
-					}
-				}
-			}
+			// sync in background
+			syncLogsAt(logId)
+				.then(() => {
+					log.info("logs synced:", logId);
+				})
+				.catch((error) => {
+					log.error(error);
+				});
 
-			return `log files(${logFiles.length}) has been uploaded to S3 logs/${deviceId}`;
+			return logId;
 		} catch (error) {
 			log.error(error);
 			return error.message;
